requirements.txt
alembic==1.5.8
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.6
Werkzeug==1.0.1
